backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
boto3==1.34.34
langchain==0.1.5
anthropic==0.18.1
python-dotenv==1.0.1
alembic==1.13.1
pytest==8.0.0
httpx==0.26.0 